refactor(xai): share tool auth fallback helpers
